requirements.txt
# Web框架
fastapi
uvicorn[standard]

# 数据库
sqlalchemy
psycopg2-binary
alembic

# Redis
redis

# 认证和安全
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# 数据验证
pydantic
pydantic-settings

# HTTP客户端
httpx
requests

# LLM集成
langchain
langchain-openai
langchain-google-genai
langchain-community
openai
google-generativeai

# 工具库
python-dotenv
typing-extensions

# 开发工具
pytest
pytest-asyncio
black
flake8
mypy